Further interface improvements and initial core data foundations.
The interface has been further refined.

Core data is now in the program with the foundation down
to do something with it! There are two windows, an overview
and a detail view. These need to be linked together. The detail
view needs work to display all the downloaded information.

Also am no longer tracking anything in the debug directory. A
.gitignore file has been created for this purpose also.