Significant work towards completing basic core-data functionality.
The detail window now has working fields for nearly all data returned by
isbndb.org.

The add and remove buttons work to a degree. Add displays the detail
window but the contents are not saved back.

The remove button removes the item from the object graph.

Also a search box was added to the toolbar.