Basic core-data functionality now complete.
The add and remove buttons both work now. The add displays
the detail window where you can search by isbn. You can then click
save and it will be saved back to the tableview.

The remove button displays a prompt confirming before removing.

Double clicking on an entry displays the the detail window with fields
loaded allowing the user to edit them and save them back to the
'database'.