Implemented auto-save.
When an action that affects the database is performed
the database is automatically saved. This is done in code
and any future actions need to save themselves.

As a result of this, pressing cancel on adding a new book
now performs a 'rollback' on the database and removes the
blank (unsaved) book.