Information is now not lost on tab change in detail view.
When the tab is changed the managed object is updated from
the UI and then the UI is updated from the managed object
causing the information not to be lost. If save is clicked
then the object is saved as usual, if cancel is clicked then
the managed object context is rollbacked anyway to get rid of
potential blank books.

Also the tabs have been reworked to include a future summary
and ownership view.