Added an icon and summary of number of books.
A first draft icon has now been added and is used by
the application.

The number of books is displayed at the bottom of the
overview window and is updated automatically when adding
or removing a book. The logic for how this is saved has
now been changed and is based on adding or removing the
object from the array controller. This should allow for
more flexibility in the future.