Reworked the book adding/saving logic.
This now fixes all of the bugs in the last few commits.
Each fix added a new bug previously. Now a delegate is
implemented and this calls a method in SofiaApplication.m
when save is clicked. This is then responsible for updating
the summary text and adding the object to the arrayController.
The BooksWindowController class saves the objectContext.

Due to this the summary works and you don't get blank books in
the library.

In order to get the summary to display on load KVO is used but
once the summary has been updated it is then unregistered.