Added Subjects and Authors tab (basic).
The subjects and authors tab now has two tables, one
for each. This allows basic adding and removing of either
that is saved into the managed object context based on the
book. This is not updated by a search yet. No logic is in
place for updating the UI or managedObject from these tables.