Implemented the rest of the manage subjects window.
The manage subjects window is now fully functional, in
terms that it has all the features of the manage authors
window. You can add, remove and choose different subjects
from the book detail window. You can open up book details
from the manage subjects window. Also if a duplicate subject
is detected it adds to it rather than creating a new one.
This isn't as hacked as the authors duplicate detection.