Added a contextual menu to the sidebar.
When right click on a list the option to rename
and delete the list has been implemented. These are
greyed out if you click on anything but a list.

Also passed on keys that aren't backspace to next
first repsonder in the sidebar and tableview.

Adding a book list also selects it and starts the
rename process.