Refactored the main views into controllers.
The two main views, image and table, had all of
their control logic within their cocoa/quartz
extended classes. This seemed to be the illogical
place to put this logic. Controllers were created for
these two views and their logic and bindings moved
here.

TODO: A class as the root of these two controllers would
seem to make sense as they will eventually share a
lot of similar logic. This should also be done for
the sidebar extension.