Got a basic import books panel working.
You can copy and paste text, it will automagically
extract ISBNs. You can also enter an URL that it will
download and extract for you. Need to hook up the buttons,
get the actual importing working and improve the regex
for extracting as a typical amazon page gets a _lot_ of
erroneous matches!