Reflection image view now displays both image and reflection.
Previously two image views were used, one displaying
the image and one displaying the reflection. They had
to be perfectly aligned. This caused (rare) problems with
the way cocoa scales images and the way I do; not lining
up. Now the image and reflection are drawn in the same
view. This keeps things neater and solves any alignment
issues.