Added method to extract TOC from amazon for a book.
The algorithm for extracting the table of contents
from amazon is a little bit of a hack to say the least.
This information isn't exposed by the API so I use the
API to get a link to the page and from there extract
it using regexps and the html.