Fixed out of bounds error. No tooltips in tableview.
Fixed out of bounds error by returning the count
of how many similar books have actually downloaded
for the number of table rows, rather than how many
will be downloaded.

Subclassed NSTextFieldCell and returned an
NSZeroRect for the expansion frame so as not to
display, what looks like but aren't, tooltips.