Added classes for coverflow and setup app in preperation.
Made changes to the xib files and SofiaApplication
to handle a coverflow view. The basics of which
have been setup.

Added a bunch of classes in preparation for
creating a custom coverflow view. This is
necessary as no public api exists for coverflow
and I couldn't find a decent open source
alternative. Time to roll my own!