Transform is now applied to images not in focus.
The images that are not focused are transformed so
that they are given a perspective effect.

Also made the distance from the focused image to
the image either side a constant function of the
maximum height. This stops the images moving
around when a very wide or very thin book is
brought into focus.