Improved perspective transform.
The perspective transform is now based on an angle
rather than the image height. This gives a much
improved aesthetic appearance when two books of
significant different width are next to each
other.

The new transform is not quite correct. The angle
that the image is tapered from the top does not
equal the angle at the bottom. This is impossible
to detect with this shallow angle. Possibly
investigate this in the future?