Added animation timing.
Added a custom cubic Bezier curve for the
animation timing. This has a sharp steep initial
rise and then plateaus for a large portion of the
remaining time.

Increased the spacing between the focused item and
surrounding items. This adds to the effect of the
animation and increases the focus of the focused
item.

Also added methods that allow programmatic
selection of the focused item as well as querying
the index of the selected item.