Added the drawing logic of the scrollbar.
Added a scroll bar that is drawn to the screen
based on the screen width. A bubble, sized
relative to the scroll bar width and number of
items, is drawn in the correct position depending
on the selected item.

Changed the positioning logic to accommodate a
scroll bar. It is no longer relative to the view
height but instead uses fixed heights.

Occasionally on startup the text and scroll bar
are not drawn. This bug needs to be investigated
and fixed.