Scrolling now possible by dragging scroll bubble.
By clicking on (only) the scroll bubble and
dragging you can scroll through the items. I had
to change the animation timing to a much faster
and linear animation for the bubble to seem
responsive. This implementation leaves the bubble
as close to the cursor as possible but is not
perfect.

Also, refactored some of the event handling logic.