Removed rollback on pressing cancel - detail window.
When cancel was pressed in the detail window a
rollback message was sent to the managed object
context. This caused any changes made globally to
be rolled back and was capable of losing a lot of
unsaved information. Now rollback is removed but
it is possible to edit the backing database and
not have the changes reversed. This is less
dangerous.