Drag and drop a new image is now possible.
It is now possible to drag and drop a new image
on to the summary view image to replace it. This
only works for the summary view as this is what is
used to update managed object.

Also added a default image for the summary
image view and cover tab for when artwork is
missing.

Updating the artwork does not work for coverflow
and image browser if they are not the currently
selected view. This needs to be fixed. Possibly
some sort of observing?