Changed noOfCopies from a string to a decimal.
Changed the noOfCopies attribute on the book
entity of the core data model to a decimal rather
than a string so that more helpful predicates
could be used.

This required creating a custom migration model to
handle the conversion. It also required a new
database model version to be created.

Testing of this migration needs to be conducted
both from version 2 and from the older version 1.