Fixed bug when animating filter notification.
If the filter notification was hidden and then
revealed too quickly (i.e programmatically) it
would try to animate whilst being animated and not
reveal; instead it would just hide.

Now checks to see if it is being animated and
schedules the next animation to start once the
current one has finished.