Removed fetching entire core data database on start up.
This was in place to ensure updating summary text
was correct. It did however load the entire
database into memory on start up. This is now no
longer the case and is handled using a
countForFetchRequest.

This allows Sofia to start much faster. The data
does however lag a little behind the app now.