Improved how Coverflow assigns images to layers.
In order to improve memory usage the Coverflow
only adds images to layers when they are to be
shown. When they are being animated off-screen it
removes the image.

This has had very little impact on the amount of
memory the app uses. It appears that Core Data
still retains the images in memory as they are
part of the managed book objects. This change does
drastically improve start up and response time
however.

Also changed the Debug database to be SQL so that
it can be much larger for efficient testing.