Reimplemented vignetting in the Coverflow.
Instead of every layer having a black sublayer
whose opacity is increased as it gets further from
the focused layer, a global vignette layer is
used. This is much more memory and computation
sensitive and as such gives better scrolling
performance. It also eliminates the 'black out'
effect (currently, only when all contents are
assigned -- to be fixed next).

Completely removed the old way of performing the
vignette. To ensure the vignette layer doesn't
respond to mouse clicks it inherits from
CAGradientLayer and overrides containsPoint;
always returning NO.