Added a double buffer to coverflow display.
Due to (commit 0b7e41c85) how the content of
coverflow layers is now only assigned when they
are to be shown; scrolling fast could produce
'holes' where layers should be drawn. The
coverflow now assigns images to layers just off
screen too, essentially creating a double buffer.
This all but eliminates the unwanted effect.

If scrolling _really_ fast, a grey background is
now used (instead of none), this at least
represents that a layer is there and is much less
unsightly.