Removed lazy fetching from array controller.
This was introduced in commit ff34b045c. This was
done as it increased start up time of Sofia. The
image data was held as part of the book managed
object and this meant that it didn't all have to
be read into memory first. With the new version of
the database this is no longer necessary.

Changed back as it introduced a bug where on
startup the library predicate wasn't being applied as it
got there before the books did! Essentially this
meant that shopping list books were being grouped
into library books.